Fix typo and add license warning
We learn from the internet that the font Black Chancery can be used
freely for commercial purpose. I did try to find the orignal source
of the statement, and the contact information of the author. However,
I had no success. I hope some guys can provide us the correct info.